+++ jest.config.js
@@ -14,12 +14,14 @@ module.exports = {
   // clearMocks: false,
 
   // Indicates whether the coverage information should be collected while executing the test
-  collectCoverage: true,
+  collectCoverage: false,
 
   // The directory where Jest should output its coverage files
   coverageDirectory: "coverage",
   // Indicates which provider should be used to instrument code for coverage
   coverageProvider: "v8",
+
+  // coveragePathIgnorePatterns: ["/node_modules/", "tests/"],
 
   // The test environment that will be used for testing
   testEnvironment: "node",