+++ tests/integration/withPage/pages/_document.js
@@ -1,0 +1,15 @@
+import Document, { Html, Main, NextScript } from 'next/document'
+export default class MyDocument extends Document {
+  render() {
+
+    return (
+      <Html lang="en">
+        <body>
+          <Main />
+          <NextScript />
+        </body>
+      </Html>
+    )
+  }
+}
+